+++ addOns/wappalyzer/CHANGELOG.md
@@ -6,8 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Updated with upstream Wappalyzer icon and pattern changes.
-
-
+- Now using 2.10 logging infrastructure (Log4j 2.x).
 
 ## [21.0.0] - 2020-12-15
 ### Changed